restaurant-qr/ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
joblib==1.3.2
textblob==0.17.1
nltk==3.8.1
scipy==1.11.4
requests==2.31.0
